// client/src/app/shared/project.model.ts
export class Project {
    ID:number;
    Title: string ;
    ProjectNumber: string ;
    ProjectStatusID: number ;
    ProjectStatusName: string ;
    ProjectPriorityID: number ;
    ProjectPriorityTitle: string ;
    PlanName: string ;
    PlanNumber: string ;
    StartDate: Date ;
    DueDate: Date ;
    JobAddress: string ;
    JobCity: string ;
    JobState: string ;
    JobZipAddress: string ;
    DWGFIleLocation: string ;
    CustomerID: number ;
    CustomerName: string ;
    ReviewedByID: number ;
    ReviewedByName: string ;
    ProjectTypeID: number ;
    ProjectTypeName: string ;
    ProjectTypeNotes: string ;
    IsFoundation: boolean ;
    FoundationTypeID: number ;
    FoundationTypeName: string ;
    CrawlSpaceJoist: string ;
    IsHaveSoilsReport: boolean ;
    How_WhenSending: string ;
    IsFraming: boolean ;
    TypeOfFloorJoist: string ;
    MaxFloorJoistSpacing: string ;
    TypeOfCeilingJoist: string ;
    RoofingMaterials: string ;
    WeightOfRoofingMaterials: string ;
    ProjectMaterialNotes: string ;
    IsHaveCourierPlans: boolean ;
    CouriertoWhom: string ;
    CouriertoAddress: string ;
    IsHaveEmail: boolean ;
    EmailAddress: string ;
    IsCustomerWillPickup: boolean ;
    PickupByPersonName: string ;
    PickupByPersonPhoneNumber: string ;
    NumberOfCopies: string ;
    ProjectHoldNotes: string ;    
    Invoice: string;
    CustomerData :string;
}